more Edits

+++ Company.A1.PL/Helpers/AttachmentSettings.cs
@@ -1,41 +1,27 @@
 ﻿namespace Company.A1.PL.Helpers
 {
-    public  static class DocumentSettings
+
+    public static class AttachmentSettings
     {
-         // Upload any file
-
-        public static string UploadFile(IFormFile file, string folderName) // return image name to store in db
+        // upload,delete
+        public static string UploadFile(IFormFile file, string folderName)
         {
-            // 1 Get Folder Location
-
-            //string folderPath = ""+folderName; 
-            //var folderPath=Directory.GetCurrentDirectory() +"\\wwwroot\\files\\" +folderName;
-           // "D:\.NET\MVC\Company.A1"
+            // File path :
+            // -Folder Location
+            //var folderPath = Directory.GetCurrentDirectory() + "\\\\wwwroot\\\\files\\" + folderName;
             var folderPath = Path.Combine(Directory.GetCurrentDirectory(), @"wwwroot\files", folderName);
-
-            // 2 Get FileName and Make it Unique
-
+            // 2. File Name
             var fileName = $"{Guid.NewGuid()}{file.FileName}";
-
-            // File Path
             var filePath = Path.Combine(folderPath, fileName);
             using var fileStream = new FileStream(filePath, FileMode.Create);
             file.CopyTo(fileStream);
-
             return fileName;
         }
-
-        // delete any file
 
         public static void DeleteFile(string fileName, string folderName)
         {
             var filePath = Path.Combine(Directory.GetCurrentDirectory(), @"wwwroot\files", folderName, fileName);
-            if (File.Exists(filePath))
-            {
-                File.Delete(filePath);
-            }
-
+            if (File.Exists(filePath)) File.Delete(filePath);
         }
-
     }
 }
